Fixed redis configuration, simplified testing.

+++ _build-and-test-all-cdc.sh
@@ -31,45 +31,45 @@ $DOCKER_COMPOSE up -d
 ./gradlew $GRADLE_OPTIONS :eventuate-tram-e2e-tests-jdbc-kafka:cleanTest :eventuate-tram-e2e-tests-jdbc-kafka:test
 
 
+if [[ "${DATABASE}" == "mysql" ]]; then
+    $DOCKER_COMPOSE stop cdcservice
+    $DOCKER_COMPOSE rm --force cdcservice
 
-$DOCKER_COMPOSE stop cdcservice
-$DOCKER_COMPOSE rm --force cdcservice
+    if [ -z "$SPRING_PROFILES_ACTIVE" ] ; then
+      export SPRING_PROFILES_ACTIVE=ActiveMQ
+    else
+      export SPRING_PROFILES_ACTIVE=${SPRING_PROFILES_ACTIVE},ActiveMQ
+    fi
 
-if [ -z "$SPRING_PROFILES_ACTIVE" ] ; then
-  export SPRING_PROFILES_ACTIVE=ActiveMQ
-else
-  export SPRING_PROFILES_ACTIVE=${SPRING_PROFILES_ACTIVE},ActiveMQ
+    $DOCKER_COMPOSE up -d cdcservice
+    ./wait-for-services.sh $DOCKER_HOST_IP 8099
+
+    ./gradlew $GRADLE_OPTIONS :eventuate-tram-e2e-tests-jdbc-activemq:cleanTest :eventuate-tram-e2e-tests-jdbc-activemq:test
+
+
+
+    $DOCKER_COMPOSE stop cdcservice
+    $DOCKER_COMPOSE rm --force cdcservice
+
+    export SPRING_PROFILES_ACTIVE=${SPRING_PROFILES_ACTIVE/ActiveMQ/RabbitMQ}
+
+    $DOCKER_COMPOSE up -d cdcservice
+    ./wait-for-services.sh $DOCKER_HOST_IP 8099
+
+    ./gradlew $GRADLE_OPTIONS :eventuate-tram-e2e-tests-jdbc-rabbitmq:cleanTest :eventuate-tram-e2e-tests-jdbc-rabbitmq:test
+
+
+
+    $DOCKER_COMPOSE stop cdcservice
+    $DOCKER_COMPOSE rm --force cdcservice
+
+    export SPRING_PROFILES_ACTIVE=${SPRING_PROFILES_ACTIVE/RabbitMQ/Redis}
+
+    $DOCKER_COMPOSE up -d cdcservice
+    ./wait-for-services.sh $DOCKER_HOST_IP 8099
+
+    ./gradlew $GRADLE_OPTIONS :eventuate-tram-e2e-tests-jdbc-redis:cleanTest :eventuate-tram-e2e-tests-jdbc-redis:test
 fi
-
-$DOCKER_COMPOSE up -d cdcservice
-./wait-for-services.sh $DOCKER_HOST_IP 8099
-
-./gradlew $GRADLE_OPTIONS :eventuate-tram-e2e-tests-jdbc-activemq:cleanTest :eventuate-tram-e2e-tests-jdbc-activemq:test
-
-
-
-$DOCKER_COMPOSE stop cdcservice
-$DOCKER_COMPOSE rm --force cdcservice
-
-export SPRING_PROFILES_ACTIVE=${SPRING_PROFILES_ACTIVE/ActiveMQ/RabbitMQ}
-
-$DOCKER_COMPOSE up -d cdcservice
-./wait-for-services.sh $DOCKER_HOST_IP 8099
-
-./gradlew $GRADLE_OPTIONS :eventuate-tram-e2e-tests-jdbc-rabbitmq:cleanTest :eventuate-tram-e2e-tests-jdbc-rabbitmq:test
-
-
-
-$DOCKER_COMPOSE stop cdcservice
-$DOCKER_COMPOSE rm --force cdcservice
-
-export SPRING_PROFILES_ACTIVE=${SPRING_PROFILES_ACTIVE/RabbitMQ/Redis}
-
-$DOCKER_COMPOSE up -d cdcservice
-./wait-for-services.sh $DOCKER_HOST_IP 8099
-
-./gradlew $GRADLE_OPTIONS :eventuate-tram-e2e-tests-jdbc-redis:cleanTest :eventuate-tram-e2e-tests-jdbc-redis:test
-
 
 $DOCKER_COMPOSE stop
 $DOCKER_COMPOSE rm --force -v
